Trekker SubjectsCombobox ut i egen komponent
